splitting linker.js file by extracting context
